fixed graphql query
