--- src/components/Catalog/DataRodsEx.json
@@ -1,68 +1,0 @@
-[
-    {
-        "Variable.Id": "GPM_3IMERGHH_07_precipitation",
-        "Collection.Id": "GPM_3IMERGHH.07",
-        "Variable.Name": "precipitation",
-        "Variable.Units": "mm/hr",
-        "Collection.East": 180,
-        "Collection.West": -180,
-        "Variable.Active": true,
-        "Collection.North": 90,
-        "Collection.South": -90,
-        "Variable.Keywords": [
-          "precipitation",
-          "rain",
-          "satellite",
-          "rain rate",
-          "gauge",
-          "snow",
-          "Huffman",
-          "GPM",
-          "rainfall",
-          "IMERG Final",
-          "IMERG V07",
-          "IMERG",
-          "IMERG Half-hourly"
-        ],
-        "Variable.LongName": "Multi-satellite precipitation estimate with gauge calibration - Final Run (recommended for general use)",
-        "Collection.Version": "07",
-        "Collection.LongName": "GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V07 (GPM_3IMERGHH) at GES DISC",
-        "Variable.Discipline": [
-          "Hydrology"
-        ],
-        "Collection.DataSetId": "GPM IMERG Final Precipitation L3 Half Hourly 0.1 degree x 0.1 degree V07 (GPM_3IMERGHH) at GES DISC",
-        "Collection.ShortName": "GPM_3IMERGHH",
-        "Variable.Measurement": [
-          "Precipitation"
-        ],
-        "Collection.DataCenter": "NASA/GSFC/SED/ESD/GCDC/GESDISC",
-        "Collection.OpendapUrl": "https://gpm1.gesdisc.eosdis.nasa.gov/opendap/GPM_L3/GPM_3IMERGHH.07/2000/153/3B-HHR.MS.MRG.3IMERG.20000601-S000000-E002959.0000.V07B.HDF5.html",
-        "Variable.StandardName": "",
-        "Collection.EndDateTime": "2023-12-31T23:59:59Z",
-        "Collection.Observation": "Observation",
-        "Collection.TimeInterval": "half-hourly",
-        "Variable.DescriptionUrl": "https://disc.gsfc.nasa.gov/information/glossary?keywords=giovanni%20measurements&title=Giovanni%20Measurement%20Definitions%3A%20Precipitation",
-        "Variable.TimeIntvRepPos": "start",
-        "Collection.BeginDateTime": "2000-06-01T00:00:00Z",
-        "Collection.EndTimeOffset": 0,
-        "Collection.TimeFrequency": 1,
-        "Collection.DescriptionUrl": "https://disc.gsfc.nasa.gov/datasets/GPM_3IMERGHH_07/summary",
-        "Variable.DestinationUnits": [
-          "inch/hr"
-        ],
-        "Collection.ProcessingLevel": "3",
-        "Collection.StartTimeOffset": 1,
-        "Collection.PlatformLongName": "Integrated Multi-satellitE Retrievals for GPM",
-        "Variable.ValuesDistribution": "linear",
-        "Collection.PlatformShortName": "IMERG",
-        "Collection.SpatialResolution": "0.1 deg.",
-        "Collection.InstrumentLongName": "Integrated Multi-satellitE Retrievals for GPM",
-        "Collection.PlatformInstrument": [
-          "GPM"
-        ],
-        "Collection.InstrumentShortName": "IMERG",
-        "Collection.SpatialResolutionUnits": "deg.",
-        "Collection.SpatialResolutionLatitude": 0.1,
-        "Collection.SpatialResolutionLongitude": 0.1
-      }
-]
