+++ src/components/Catalog/Variables.tsx
@@ -5,7 +5,6 @@ import {
   IonItem,
   IonLabel,
   IonList,
-  IonListHeader,
 } from "@ionic/react";
 
 import catalog from "./catalog.json";
@@ -14,26 +13,9 @@ interface VariablesProps {
   onVariableChange: (value: string) => void;
 }
 
-// const topics = [
-//   { group: "Atmosphere", subgroups: ["Surface", "Atmospheric Composition"] },
-//   { group: "Land", subgroup: ["Hydrology", "Cryosphere", "Biology"] },
-//   { group: "Ocean", subgroup: "" },
-// ];
 const topics = ["Atmosphere", "Land", "Ocean"];
 
 const Variables: React.FC<VariablesProps> = ({ onVariableChange }) => {
-  // const [expanded, setExpanded] = useState<string>("");
-  const handleClick = (value: string) => {
-    onVariableChange(value);
-  };
-
-  //   const handleItemClick = (value: string) => {
-  //     setExpanded(expanded === value ? "" : value);
-  //     if (value === "precipitation") {
-  //       handleClick();
-  //     }
-  //   };
-
   const displayCatalog = topics.map((topic) => {
     return (
       <IonAccordion key={topic} value={topic}>
@@ -47,7 +29,7 @@ const Variables: React.FC<VariablesProps> = ({ onVariableChange }) => {
               return (
                 <IonItem
                   button
-                  onClick={handleClick.bind(null, data.dataFieldId)}
+                  onClick={() => onVariableChange(data.dataFieldId)}
                   key={data.label}
                   color="light"
                 >
