+++ src/services/api/time-series-exp.ts
@@ -1,39 +1,12 @@
 
-// import { format } from 'date-fns'
+import { parseTimeSeriesCsv } from '../../utils/time-series';
 
 import type {
     MaybeBearerToken,
-    TimeSeriesData,
-    TimeSeriesDataRow,
-    TimeSeriesMetadata,
     DataParams,
 } from './time-series.types'
 
-const parseTimeSeriesCsv = (text: string) => {
-    const lines = text.split('\n')
 
-    // get dates only
-    const dateLines = lines.slice(lines.indexOf("")+2);
-
-    const metadata: Partial<TimeSeriesMetadata> = {}
-    const data: TimeSeriesDataRow[] = []
-
-    dateLines.forEach(line => {
-        if (line.includes('=')) {
-            const [key, value] = line.split('=')
-            metadata[key] = value
-        } else if (line.includes(',')) {
-            const [timestamp, value] = line.split(',')
-            if (timestamp && value) {
-                data.push({ timestamp, value })
-            }
-        }
-    })
-    console.log("metadata: ", metadata)
-    console.log("data: ", data)
-
-    return { metadata, data } as TimeSeriesData;
-}
 
 export const fetchData = async (dataParams: DataParams) => {
     const bearerToken: MaybeBearerToken = null;
@@ -68,3 +41,47 @@ export const fetchData = async (dataParams: DataParams) => {
     }
 }
 
+/**
+ * fetching data from graphql
+ */
+// pass variable id
+// {"query":"{\n  getVariables(variableEntryIds: [\"OMAERUVd_003_FinalAerosolAbsOpticalDepth388\"]) { variables { dataFieldId, dataFieldLongName } } }"}
+// export const fetchCatalog = async () => {
+//     const url =
+//       "https://u2u5qu332rhmxpiazjcqz6gkdm.appsync-api.us-east-1.amazonaws.com/graphql";
+//     const query = `
+//                 query {
+//                     getVariables {
+//                     variables {
+//                         dataFieldId
+//                         dataFieldLongName
+//                     }
+//                     }
+//                 }
+//                 `;
+//     const requestOptions = {
+//       method: "POST",
+//       headers: {
+//         "Content-Type": "application/json",
+//         "x-api-key": "",
+//       },
+//       // body: "query":"{\n  getVariables { variables { dataFieldId, dataFieldLongName } } }",
+//       body: JSON.stringify({ query }),
+//     };
+
+//     try {
+//       console.log("loading catalog data...");
+
+//       const response = await fetch(
+//         url,
+//         requestOptions
+//       );
+//       // const response = await axios.post(url, requestOptions);
+//       // const data = await response.json();
+//       console.log(response);
+//     } catch (error) {
+//       console.log("Catalog error: ", error);
+//     } finally {
+//       // console.log("catalog data loaded...");
+//     }
+//   };
